move expr updates to separate actor
Splits out the expression update to a separate actor and
switches the publish actor to use a round robin dispatcher.
