+++ iep-lwc-bridge/src/main/resources/application.conf
@@ -17,6 +17,10 @@ atlas {
       {
         name = "publish"
         class = "com.netflix.iep.lwc.LwcPublishActor"
+      },
+      {
+        name = "lwc-expr-update"
+        class = "com.netflix.iep.lwc.ExprUpdateActor"
       }
     ]
   }
@@ -24,6 +28,13 @@ atlas {
   webapi.publish {
     // This is just a pass through, do not intern the values...
     intern-while-parsing = false
+  }
+}
+
+akka.actor.deployment {
+  /publish {
+    router = round-robin-pool
+    nr-of-instances = 8
   }
 }
